api/pyproject.toml
[tool.poetry]
name = "Docker API 101"
version = "0.1.0"
description = "Docker + FastAPI + SQLModel + ReactJS Modern API Example"
authors = ["Kyosuke Miyamura <ask@386.jp>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"

pytest = "^6.2.5"
python-dotenv = "^0.19.2"

fastapi = "^0.68.1"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}

sqlmodel = "^0.0.5"
psycopg2-binary = "^2.9.1"
alembic = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
